+++ ConsoleInvaders/console_screen.h
@@ -15,7 +15,7 @@ class console_screen
 	void construct_screen();
 	void construct_buffer();
 
-	void clip(float& x, float& y) const;
+	void clip(int& x, int& y) const;
 
 public:
 	console_screen(int width, int height, const wchar_t* title);
@@ -28,8 +28,8 @@ public:
 	void flip() const;
 	void present() const;
 
-	void plot_char(float x, float y, wchar_t chr) const;
-	void draw_text(float x, float y, size_t size, const wchar_t* text) const;
-	void draw_text_centered(float y, size_t size, const wchar_t* text) const;
+	void plot_char(int x, int y, wchar_t chr) const;
+	void draw_text(int x, int y, size_t size, const wchar_t* text) const;
+	void draw_text_centered(int y, size_t size, const wchar_t* text) const;
 };
